refactor!: align `issue` signature with `anchorCredential`
